Properly added capmonster support
